Fix LP64 allocache pointer cast in sound loading
Replace (long *)&Sound[num].ptr with (intptr_t *)&Sound[num].ptr in
PREMAP.C and SOUNDS.C. Sound[num].ptr is char*, so &Sound[num].ptr is
char**. Casting to (long *) is semantically incorrect — allocache writes
a pointer-sized value through the handle. On LP64 (sizeof(long) ==
sizeof(char*) == 8) this works by accident, but (intptr_t *) correctly
expresses the intent of storing a pointer-sized cache address.

+++ source/PREMAP.C
@@ -250,7 +250,7 @@ char getsound(unsigned short num)
         ( l < 12288 ) )
     {
         Sound[num].lock = 2;
-        allocache((long *)&Sound[num].ptr,l,&Sound[num].lock);
+        allocache((intptr_t *)&Sound[num].ptr,l,&Sound[num].lock);
         if(Sound[num].ptr != NULL)
             kread( fp, Sound[num].ptr , l);
     }
