+++ source/SOUNDS.C
@@ -298,7 +298,7 @@ char loadsound(unsigned short num)
 
     Sound[num].lock = 200;
 
-    allocache((long *)&Sound[num].ptr,l,&Sound[num].lock);
+    allocache((intptr_t *)&Sound[num].ptr,l,&Sound[num].lock);
     kread( fp, Sound[num].ptr , l);
     kclose( fp );
     return 1;
